+++ RandomEvents/EventSkillsOnly.cs
@@ -27,8 +27,8 @@ namespace RandomEvents
         }
         public override bool Condition(List<AbstractEvent> activeOtherEvents)
         {
-            var equipmentOnlyEventActive = activeOtherEvents.Any(e => e.GetEventConfigName().Equals("EquipmentOnly", StringComparison.InvariantCultureIgnoreCase));
-
+            var equipmentOnlyOrMonsterTransformActive = activeOtherEvents.Any(e => e.GetEventConfigName().Equals("EquipmentOnly", StringComparison.InvariantCultureIgnoreCase) ||
+                                                                                   e.GetEventConfigName().Equals("MonsterTransform", StringComparison.InvariantCultureIgnoreCase));
             // none of the characters must have the Bandolier item, as it breaks the concept of this event
             var hasBandolier = false;
             foreach (PlayerCharacterMasterController pc in PlayerCharacterMasterController.instances)
@@ -43,7 +43,7 @@ namespace RandomEvents
                 }
             }
 
-            return !equipmentOnlyEventActive && !hasBandolier;
+            return !equipmentOnlyOrMonsterTransformActive && !hasBandolier;
         }
         public override string GetEventConfigName()
         {
@@ -59,7 +59,7 @@ namespace RandomEvents
         }
         public override string GetConditionDescription()
         {
-            return "None of the players have a bandolier in their inventories (workaround to make it work server-side only).";
+            return "Events \"EquipmentOnly\", \"MonsterTransform\" inactive and none of the players have a bandolier in their inventories (workaround to make it work server-side only).";
         }
         protected override void AddConfig(ConfigFile config)
         {