+++ RandomEvents/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+- Update Icon
+- Add `MonsterTransform` Event - players turn into monsters
+- Fix `BulletHell` to actually increase Fire Rate of monsters
+- Prevent `Aritfact of Enigma` and `EquipmentOnly` from occuring at the same time
+- Fix event conlusion message appearing even if no event is active
+
 # 1.0.1
 
 - Improve Icon